refactor: create new component Switchers, moved the language and theme switches to separate components
