+++ components/header/Switchers.jsx
@@ -1,20 +1,22 @@
 "use client";
 import { useState } from "react";
-import Link from "next/link";
+import ThemeSwitcher from "./ThemeSwitcher";
+import LanguageSwitcher from "./LanguageSwitcher";
 import Image from "next/image";
 
-const LanguageSwitcher = () => {
+const DropdownSwitcher = ({ content, options }) => {
   const [isDropdownVisible, setIsDropdownVisible] = useState(false);
 
   const toggleDropdown = () => {
     setIsDropdownVisible(!isDropdownVisible);
   };
+
   return (
     <div
-      className=" dropdown-block relative w-[89px] px-4 pb-1"
+      className="dropdown-block relative w-[89px] px-4 pb-1"
       onClick={toggleDropdown}
     >
-      <span className="title-switcher relative cursor-pointer">Мова</span>
+      <span className="title-switcher relative cursor-pointer">{content}</span>
       <Image
         src="/icons/arrow.svg"
         width="20"
@@ -23,23 +25,23 @@ const LanguageSwitcher = () => {
           isDropdownVisible ? "transform rotate-180" : ""
         }`}
       />
-      <ul
+      <div
         className={`dropdown w-full absolute top-7 left-0 flex flex-col gap-[10px] border rounded border-solid border-gray p-4 transition-opacity duration-300 
-       ${isDropdownVisible ? "visible opacity-100" : "invisible opacity-0"} `}
+         ${isDropdownVisible ? "visible opacity-100" : "invisible opacity-0"} `}
       >
-        <li className="text-center h-5">
-          <Link href="/" style={{ height: "100%" }}>
-            Укр
-          </Link>
-        </li>
-        <li className="text-center h-5">
-          <Link href="/En" style={{ height: "100%" }}>
-            Eng
-          </Link>
-        </li>
-      </ul>
+        {options}
+      </div>
     </div>
   );
 };
 
-export default LanguageSwitcher;
+const Switchers = () => {
+  return (
+    <div className="flex">
+      <DropdownSwitcher content="Тема" options={<ThemeSwitcher />} />
+      <DropdownSwitcher content="Мова" options={<LanguageSwitcher />} />
+    </div>
+  );
+};
+
+export default Switchers;
