+++ components/header/ThemeSwitcher.jsx
@@ -1,0 +1,14 @@
+const ThemeSwitcher = () => {
+  return (
+    <>
+      <button id="dark-mode" type="button" className="text-center h-5">
+        Темна
+      </button>
+      <button id="light-mode" type="button" className="text-center h-5">
+        Світла
+      </button>
+    </>
+  );
+};
+
+export default ThemeSwitcher;
